fix linux compile errors

+++ nanojit/LIR.cpp
@@ -1046,7 +1046,7 @@ namespace nanojit
         NanoAssert(j == argc);
 #endif
 
-		NanoAssert(argc <= MAXARGS);
+		NanoAssert(argc <= (int)MAXARGS);
 		uint32_t words = argwords(argc);
 		ensureRoom(words+argc+1);  // ins size + possible tramps
 		for (int32_t i=0; i < argc; i++)
